Add Insert Amount Cargo

+++ app/src/main/java/com/example/gdemobile/models/DocumentPosition.kt
@@ -1,5 +1,5 @@
 package com.example.gdemobile.models
 
-class DocumentPosition (code :  String,  name : String,unit: String,   barcode: String, var amount : Int) : Cargo(code,name,barcode,unit) {
+class DocumentPosition (code :  String,  name : String,unit: String,   barcode: String, var amount : Double) : Cargo(code,name,barcode,unit) {
 
 }
